[JSC] Collect execution count in Wasm::BaselineData https://bugs.webkit.org/show_bug.cgi?id=301886 rdar://163966362
Reviewed by Yijia Huang.

We were using Callee's execution counter to collect total execution
count of the function in the inlining decision. But this is too fragile
and unstable as BBQCallee can go away. And we found some cases these
data is missing since OMG for that callee is already compiled and
installed. Let's collect it in Wasm::BaselineData, this is much more
consistent since the other CallProfile data is also obtained from
Wasm::BaselineData.

* Source/JavaScriptCore/llint/InPlaceInterpreter.asm:
* Source/JavaScriptCore/wasm/WasmBBQJIT.cpp:
(JSC::Wasm::BBQJITImpl::BBQJIT::emitIncrementCallProfileCount):
(JSC::Wasm::BBQJITImpl::BBQJIT::addTopLevel):
(JSC::Wasm::BBQJITImpl::BBQJIT::addLoopOSREntrypoint):
* Source/JavaScriptCore/wasm/WasmBaselineData.h:
* Source/JavaScriptCore/wasm/WasmCallee.cpp:
(JSC::Wasm::IPIntCallee::needsProfiling const): Deleted.
* Source/JavaScriptCore/wasm/WasmCallee.h:
* Source/JavaScriptCore/wasm/WasmCalleeGroup.cpp:
(JSC::Wasm::CalleeGroup::tryGetBBQCallee): Deleted.
* Source/JavaScriptCore/wasm/WasmCalleeGroup.h:
* Source/JavaScriptCore/wasm/WasmIPIntSlowPaths.cpp:
(JSC::IPInt::WASM_IPINT_EXTERN_CPP_DECL):
* Source/JavaScriptCore/wasm/WasmIPIntSlowPaths.h:
* Source/JavaScriptCore/wasm/WasmInliningDecision.cpp:
(JSC::Wasm::InliningNode::inlineNode):
* Source/JavaScriptCore/wasm/WasmMergedProfile.cpp:
(JSC::Wasm::MergedProfile::MergedProfile):
(JSC::Wasm::MergedProfile::merge):
* Source/JavaScriptCore/wasm/WasmMergedProfile.h:
(JSC::Wasm::MergedProfile::totalCount const):
* Source/JavaScriptCore/wasm/WasmModule.cpp:
(JSC::Wasm::Module::createMergedProfile):

Canonical link: https://commits.webkit.org/302505@main

+++ Source/JavaScriptCore/wasm/WasmInliningDecision.cpp
@@ -147,7 +147,7 @@ void InliningNode::inlineNode(InliningDecision& decision)
 
             double relativeCallCount = 0;
             if (profile->totalCount())
-                relativeCallCount = callCount / profile->totalCount();
+                relativeCallCount = callCount / static_cast<double>(profile->totalCount());
             size_t wasmSize = decision.m_module.moduleInformation().functionWasmSizeImportSpace(candidateCallee->index());
             auto& child = decision.m_arena.alloc(target, this, callSite.size(), index, wasmSize, relativeCallCount);
             callSite.append(&child);
